refactor: rename e2e test.
